Copied and ran patch-seq-data-load
Ran patch-seq-data-load.ipynb from the original project. I didn't run into any problems with the filepaths and URLs this time, so the file is not modified from the original.

This script generates one file, m1.pickle, but it is not tracked so it will not show up in the repo.