Ran ttype-assignment
Ran the last notebook for preprocessing, ttype-assignment.ipynb. This creates multiple images in the figures directory. It also saves a pickle file that is not tracked in this repo.

I also created a figures directory so that ttype-assignment.ipynb can write images. This directory is not directly copied from the original work, and I just made it on the fly so the code can run.