Changed name and executed ttype-coverage
Added descriptions and commented out a line that no longer works in ttype-coverage.ipynb. The reason for commenting out the line can be found in the notebook. Changed the file name because it's no longer the original.

The commented out line was like this:
col = clusterColors[clusterNames=='L5 ET_1'][0] 

Ran the file to see if it works, and in the process created the "extra" directory and "mainfigures" directory in figures because the notebook writes 
images there.