Added AMI and FMS scores to the evaluation
evaluate_ami_fms() in kNN_evaluation.py is a new helper function that calculated and plots the adjusted mutual information and Fowlkes-Mallows score. Modified revisit-confusion-matrices.ipynb so that it can show those evaluation metrics with the confusion matrix.

Also, changed the way transcriptomic types are assigned in the confusion matrix calculation. In the previous version, no type assignments were made if the majority vote indicated nan, but now type assignments are made based on the next highest vote.