Added feature processing module
The feature dictionary Scala et al. created was not exactly what I wanted to use to evaluate kNN, so I needed to create another dictionary myself. The code for it will be long, so I decided to store  it in features.py.

Right now features.py and revisit-confusion-matrices.ipynb both have code that processes the transcriptomic features.